Auto deploy CloudFormation stack
Auto deploy a AWS CloudFormation stack and start making the project self
describing as Infra-as-Code. This initial commit sets the ground work
for later resources to be added to make the template complete.

Part of this groundwork is to update AWS CLI from v1 to v2 to enable
modern features, as well as preventing dependency on legacy systems.